Implicit conversion loses integer precision: 'unsigned long' to 'int32_t'

+++ mz_crypt_openssl.c
@@ -65,7 +65,7 @@ typedef struct mz_crypt_sha_s {
     EVP_MD_CTX     *ctx;
 #endif
     int32_t        initialized;
-    int32_t        error;
+    unsigned long  error;
     uint16_t       algorithm;
 } mz_crypt_sha;
 
@@ -270,7 +270,7 @@ void mz_crypt_sha_delete(void **handle) {
 
 typedef struct mz_crypt_aes_s {
     int32_t    mode;
-    int32_t    error;
+    unsigned long    error;
     EVP_CIPHER_CTX *ctx;
 } mz_crypt_aes;
 
@@ -494,7 +494,7 @@ typedef struct mz_crypt_hmac_s {
     EVP_MAC_CTX *ctx;
 #endif
     int32_t     initialized;
-    int32_t     error;
+    unsigned long     error;
     uint16_t    algorithm;
 } mz_crypt_hmac;
